Some code refactoring

+++ rpmjukebox/src/main/java/uk/co/mpcontracting/rpmjukebox/component/PlaylistListCellFactory.java
@@ -1,16 +1,11 @@
 package uk.co.mpcontracting.rpmjukebox.component;
 
-import javafx.event.ActionEvent;
-import javafx.event.EventHandler;
 import javafx.scene.control.ContextMenu;
 import javafx.scene.control.ListCell;
 import javafx.scene.control.ListView;
 import javafx.scene.control.MenuItem;
-import javafx.scene.input.ContextMenuEvent;
-import javafx.scene.input.DragEvent;
 import javafx.scene.input.Dragboard;
 import javafx.scene.input.MouseButton;
-import javafx.scene.input.MouseEvent;
 import javafx.scene.input.TransferMode;
 import javafx.util.Callback;
 import uk.co.mpcontracting.ioc.ApplicationContext;
@@ -41,39 +36,36 @@ public class PlaylistListCellFactory extends EventAwareObject implements Callbac
 		// Mouse Events //
 		//////////////////
 
-		listCell.setOnMouseClicked(new EventHandler<MouseEvent>() {
-			@Override
-			public void handle(MouseEvent event) {
-				if (event.getButton() == MouseButton.PRIMARY) {
-					// This doesn't work as the default cell behaviour class WAY down the
-					// stack interprets a double click as a request to edit and doesn't
-					// forward the event onwards
-					/*if (event.getClickCount() > 1) {
-						// Double click
-						if (listCell != null && listCell.getItem() != null) {
-							playlistManager.playPlaylist(listCell.getItem().getPlaylistId());
-						}
-					} else {
-						// Single click
-						if (listCell != null && listCell.getItem() != null) {
-							fireEvent(Event.PLAYLIST_SELECTED, listCell.getItem().getPlaylistId());
-						}
-					}*/
-					if (event.getClickCount() == 1) {
-						// Single click
-						if (listCell != null && listCell.getItem() != null) {
-							fireEvent(Event.PLAYLIST_SELECTED, listCell.getItem().getPlaylistId());
-						}
+		listCell.setOnMouseClicked(event -> {
+			if (event.getButton() == MouseButton.PRIMARY) {
+				// This doesn't work as the default cell behaviour class WAY down the
+				// stack interprets a double click as a request to edit and doesn't
+				// forward the event onwards
+				/*if (event.getClickCount() > 1) {
+					// Double click
+					if (listCell != null && listCell.getItem() != null) {
+						playlistManager.playPlaylist(listCell.getItem().getPlaylistId());
 					}
-				} else if (event.getButton() == MouseButton.SECONDARY && event.getClickCount() < 2) {
-					// Right click
+				} else {
+					// Single click
+					if (listCell != null && listCell.getItem() != null) {
+						fireEvent(Event.PLAYLIST_SELECTED, listCell.getItem().getPlaylistId());
+					}
+				}*/
+				if (event.getClickCount() == 1) {
+					// Single click
 					if (listCell != null && listCell.getItem() != null) {
 						fireEvent(Event.PLAYLIST_SELECTED, listCell.getItem().getPlaylistId());
 					}
 				}
+			} else if (event.getButton() == MouseButton.SECONDARY && event.getClickCount() < 2) {
+				// Right click
+				if (listCell != null && listCell.getItem() != null) {
+					fireEvent(Event.PLAYLIST_SELECTED, listCell.getItem().getPlaylistId());
+				}
 			}
 		});
-		
+
 		//////////////////
 		// Context Menu //
 		//////////////////
@@ -81,60 +73,43 @@ public class PlaylistListCellFactory extends EventAwareObject implements Callbac
 		ContextMenu contextMenu = new ContextMenu();
 		
 		final MenuItem newPlaylistItem = new MenuItem(messageManager.getMessage(MESSAGE_PLAYLIST_CONTEXT_NEW_PLAYLIST));
-		newPlaylistItem.setOnAction(new EventHandler<ActionEvent>() {
-			@Override
-			public void handle(ActionEvent event) {
-				playlistManager.createPlaylist();
-			}
+		newPlaylistItem.setOnAction(event -> {
+			playlistManager.createPlaylist();
 		});
 		contextMenu.getItems().add(newPlaylistItem);
 		
 		final MenuItem deletePlaylistItem = new MenuItem(messageManager.getMessage(MESSAGE_PLAYLIST_CONTEXT_DELETE_PLAYLIST));
-		deletePlaylistItem.setOnAction(new EventHandler<ActionEvent>() {
-			@Override
-			public void handle(ActionEvent event) {
-				Playlist playlist = listView.getSelectionModel().getSelectedItem();
-				
-				ApplicationContext.getBean(MainPanelController.class).showConfirmWindow(messageManager.getMessage(MESSAGE_PLAYLIST_DELETE_ARE_YOU_SURE, playlist.getName()), 
-					true,
-					new Runnable() {
-						@Override
-						public void run() {
-							playlistManager.deletePlaylist(playlist.getPlaylistId());
-						}
-					},
-					new Runnable() {
-						@Override
-						public void run() {
-							// No-op
-						}
-					}
-				);
-			}
+		deletePlaylistItem.setOnAction(event -> {
+			Playlist playlist = listView.getSelectionModel().getSelectedItem();
+
+			ApplicationContext.getBean(MainPanelController.class).showConfirmWindow(messageManager.getMessage(MESSAGE_PLAYLIST_DELETE_ARE_YOU_SURE, playlist.getName()), 
+				true,
+				() -> {
+					playlistManager.deletePlaylist(playlist.getPlaylistId());
+				},
+				null
+			);
 		});
 		contextMenu.getItems().add(deletePlaylistItem);
 		
 		listCell.setContextMenu(contextMenu);
-		listCell.setOnContextMenuRequested(new EventHandler<ContextMenuEvent>() {
-			@Override
-			public void handle(ContextMenuEvent event) {
-				@SuppressWarnings("unchecked")
-				ListCell<Playlist> sourceCell = (ListCell<Playlist>)event.getSource();
-		
-				if (listView.getItems().size() > sourceCell.getIndex()) {
-					Playlist playlist = listView.getItems().get(sourceCell.getIndex());
-		
-					if (playlist.getPlaylistId() < 0) {
-						deletePlaylistItem.setDisable(true);
-					} else {
-						deletePlaylistItem.setDisable(false);
-					}
-		
-					newPlaylistItem.setDisable(true);
-				} else {
-					newPlaylistItem.setDisable(false);
+		listCell.setOnContextMenuRequested(event -> {
+			@SuppressWarnings("unchecked")
+			ListCell<Playlist> sourceCell = (ListCell<Playlist>)event.getSource();
+	
+			if (listView.getItems().size() > sourceCell.getIndex()) {
+				Playlist playlist = listView.getItems().get(sourceCell.getIndex());
+	
+				if (playlist.getPlaylistId() < 0) {
 					deletePlaylistItem.setDisable(true);
+				} else {
+					deletePlaylistItem.setDisable(false);
 				}
+	
+				newPlaylistItem.setDisable(true);
+			} else {
+				newPlaylistItem.setDisable(false);
+				deletePlaylistItem.setDisable(true);
 			}
 		});
 		
@@ -142,53 +117,41 @@ public class PlaylistListCellFactory extends EventAwareObject implements Callbac
 		// Drag And Drop //
 		///////////////////
 		
-		listCell.setOnDragOver(new EventHandler<DragEvent>() {
-			@Override
-			public void handle(DragEvent event) {
-				if (event.getGestureSource() != listCell && event.getDragboard().hasContent(DND_TRACK_DATA_FORMAT)) {
-					event.acceptTransferModes(TransferMode.COPY);
-				}
-		
-				event.consume();
+		listCell.setOnDragOver(event -> {
+			if (event.getGestureSource() != listCell && event.getDragboard().hasContent(DND_TRACK_DATA_FORMAT)) {
+				event.acceptTransferModes(TransferMode.COPY);
 			}
+	
+			event.consume();
 		});
-		
-		listCell.setOnDragEntered(new EventHandler<DragEvent>() {
-			@Override
-			public void handle(DragEvent event) {
-				if (event.getGestureSource() != listCell && listCell.getItem() != null && listCell.getItem().getPlaylistId() > -1 && 
-					event.getDragboard().hasContent(DND_TRACK_DATA_FORMAT)) {
-					listCell.setStyle("-fx-background-color: -jb-foreground-color; -fx-text-fill: -jb-background-color");
-				}
-		
-				event.consume();
+
+		listCell.setOnDragEntered(event -> {
+			if (event.getGestureSource() != listCell && listCell.getItem() != null && listCell.getItem().getPlaylistId() > -1 && 
+				event.getDragboard().hasContent(DND_TRACK_DATA_FORMAT)) {
+				listCell.setStyle("-fx-background-color: -jb-foreground-color; -fx-text-fill: -jb-background-color");
 			}
+	
+			event.consume();
 		});
-		
-		listCell.setOnDragExited(new EventHandler<DragEvent>() {
-			@Override
-			public void handle(DragEvent event) {
-				listCell.setStyle(null);
-		
-				event.consume();
+
+		listCell.setOnDragExited(event -> {
+			listCell.setStyle(null);
+			
+			event.consume();
+		});
+
+		listCell.setOnDragDropped(event -> {
+			Dragboard dragboard = event.getDragboard();
+			
+			if (dragboard.hasContent(DND_TRACK_DATA_FORMAT)) {
+				playlistManager.addTrackToPlaylist(listCell.getItem().getPlaylistId(), ((Track)dragboard.getContent(DND_TRACK_DATA_FORMAT)).clone());
+	
+				event.setDropCompleted(true);
 			}
+	
+			event.consume();
 		});
-		
-		listCell.setOnDragDropped(new EventHandler<DragEvent>() {
-			@Override
-			public void handle(DragEvent event) {
-				Dragboard dragboard = event.getDragboard();
-		
-				if (dragboard.hasContent(DND_TRACK_DATA_FORMAT)) {
-					playlistManager.addTrackToPlaylist(listCell.getItem().getPlaylistId(), ((Track)dragboard.getContent(DND_TRACK_DATA_FORMAT)).clone());
-		
-					event.setDropCompleted(true);
-				}
-		
-				event.consume();
-			}
-		});
-		
+
 		return listCell;
 	}
 }
